Padding hero section for mobile changed
Removed default padding and switched to 0 because white space on most phones looked excessive, which prevented users from seeing the video upon opening the home page.

Important note: it's crucial to test directly on different device with mobile internet browsers to truly see how much of a page is visible, prototyping doesn't show the native navbar on mobile for instance.